Updated TIdUDPServer.GetBinding() to not create 2 default Binding objects on Linux/Android systems.
Fix for TIdMessageClient.ReceiveBody() not parsing headers correctly when TIdMessage.IsMsgSinglePartMime is True.

No longer setting the TIdMessagePart.ExtraHeaders.NameValueSeparator property to '=' in TIdMessageClient and TIdIMAP4 when parsing emails.

Fix for TIdURI.ParamsEncode() not encoding #128.

Updated TIdFileCreateStream and TIdAppendFileStream to include the fmOpenReadWrite and fmShareDenyWrite flags when opening a file with the fmCreate flag.  This allows other processes to have read-only access to the file in Delphi 2010+, but not earlier versions where fmCreate requires exclusive access.

Added various TODO comments